requirements-ocr.txt
# OCR Worker dependencies
# Uses EasyOCR with PyTorch for better Portuguese/Latin support

# OCR engine
easyocr>=1.7.0  # Good Portuguese support, uses PyTorch

# PDF handling
PyMuPDF>=1.24.0

# Image processing
Pillow>=10.0.0
pillow-heif>=0.15.0

# API and async
aiohttp>=3.9.0
redis>=5.0.0
fastapi>=0.115.0
starlette>=0.40.0
httpx>=0.27.0

# Config and logging
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=24.0.0

# Metrics
prometheus-client>=0.19.0
